Allow empty work_dir in gimlet executable item
This is more consistent with how other project items work.
By raising ValueError at the moment of creating the executable item,
we prevented all other project items from executing.

Now, only the gimlet fails, but in the future we may find a way
to work around the empty work directory situation. At the moment,
it's difficult for the work directory to be empty, as the only case
where this happens is whenever toolbox fails to create the directory
because of permissions or something.