Get rid of the widgets package, and import it from spinetoolbox instead
The widgets package was created here to support spine_items as a
dependency of toolbox, which is no longer the case. For this reason,
we move all the submodules back to spinetoolbox.widget and import
them from there.

It makes way more sense to have all those widgets in toolbox since
more project items, even community developed ones, can use them.

Also, bump REQUIRED_SPINE_TOOLBOX_VERSION to match the version of
toolbox that has the right widgets package that we need here.