Move REQUIRED_SPINE_TOOLBOX_VERSION to version.py
This is so it can be obtained by spinetoolbox via parsin version.py
(parsing setup.py fails because it tries to run setup.py at the
same time, weird).

And get rid of the install and develop command override.
It doesn't work well with pip, notably pip install won't pass
the --install-option to setup.py.

pip has its own idiosyncrasies and it's better not to rely on it.