Change to work directory before console-executing tool instance
The workdir of the tool can change as the user switches between
using source of work dir. Moreover, if they use work dir it
changes *everytime*. So we absolutely need to cd to the right
dir before running the instance in console; it's not enough
to set the cwd argument, because the console is reused and the
cwd can become obsolete.