Add support to create Julia sysimages
Re Spine-project/Spine-Toolbox#1225

The sysimage is created through a new small new widget that's embedded
in the properties UI, whenever the user switches to a spec of Julia
type. The idea is to extend this concept with the other tool types
as needed, and the code is made for that.

The location of the julia sysimage is added to the Tool item's
dictionary and so it ends up in the project file.
Before using the image, the tool verifies that the file exist to avoid
eventual portability problems.