Add option for loading default mapping to Import editor
For connectors that support it (only excel at the moment) we
show a button that says Load default mapping. When clicked,
the workbook is examined and if it has the default excel mapping,
we load it into the editor.

The previous behavior was to load the default mapping automatically
but I think this is better, so the user has better control on what's
happening.