Avoid undo stack when creating initial mapping
Exporter's specification editor used to artificially click the 'New
mapping' button if the specification didn't exist to get users started.
This pushed an action to the undo stack -- not ideal since this would
fire an undo notification too. We now create the initial Mapping
directly without using undo stack as a middle man.