Rename ImportEditor to ImportSources
And also clarify the docstrings of ImportEditorWindow components.
I'm hoping this helps us understand the organization a bit more.
This widget grew quite organically so it's a bit arbitrary how it
works. A lots of signals being fired here and there to notify
some components of what others do. I wonder if one day we'd have
time to refactor it a little bit better.