Fix file system watcher when watching single files
Even when file system watcher is watching an individual file it still
watches the file's directory. The previous implementation didn't take
into account that in this situation other unwatched files in the
directory might be renamed, deleted, etc. We now check that a change in
a watched directory actually is about the watched file and not some
random one.