Don't write passwords to commit message
Merger was writing the entire database URL to the commit message when it
finished with its job. As URLs may contain usernames and passwords, this
was pretty dangerous operation. We now clear the URL from credentials
before writing the commit message (or any other visible log message).

Re Spine-project/Spine-Toolbox#1293