Try to fix Exporter specification editor crash
Exporter's PreviewTreeModel was setting its model index internal
pointer to point to mapping names. Apparently, sometimes the name (a
Python string object) got garbage collected in Python while Qt was
still holding the pointer in C++ side causing Tracebacks and, in the
worst case, crash of the entire application.

In this commit, we replace the pointers-to-string by mapping row
numbers, i.e. the internal pointer is interpreted as a row index + 1.
This way C++ holds just a raw int and not a pointer to something that
can be garbage collected in Python. It is highly possible that this
fixes the random crashes.