Automatically initialize empty dbs upon execution
Instead of showing a notification asking the user to do so.
In general, when we share projects in git we don't want to push
output db files, but we don't want the user to have to initialize
all the dbs manually (there can be a lot of them). This works well
with the new interpretation of the `create` argument to
`DatabaseMapping` that does not overwrite non-empty, non-spine dbs
(so it's safe).