Add scenario name to output directory names
Tool and Exporter now prepend output directory names by scenario name.
The names are filter id hashes which are not human readable so this
should make it somewhat easier to identify which directory is which
without needing to look into .filter_id files. Note that the scenario
name is only used in case there is only a single scenario involved in
filtering. This should keep the directory name length within acceptable
bounds.

Re Spine-project/Spine-Toolbox#1712