Make database purging more controllable
Importer and Merger used to purge databases by replacing the target
database by a freshly created one. There was no way to control which
database items needed purging and no trace of the operation in
database's commit log.

With this commit, Importer and Merger purge the target database the same
way as Toolbox Database editor: by removing the items using
DatabaseMapping and adding an entry to the commit log. Also, there is
now a settings dialog which can be used to select what to purge
similarly to Database editor. In fact, the relevant widgets from the
Database editor are used by also by the project items.

Re Spine-project/Spine-Toolbox#1747