Fix changing dialect in Data Store
Data store would store a broken URL if the dialect was changed from a
network database (e.g. MySQL) to sqlite because network information such
as port number could be left to the sqlite URL. We now clean network
information from the URL when the dialect gets changed to sqlite.