Fix Tool's excluded execution resources
Tool's execution item updates its _output_dir in execute() if it is
being executed filtered. This ensures that it can provide the correct
output resources downstream. However, when Tool is excluded from
execution, _output_dir is not updated. This results in
_output_resources_forward() picking some output files not necessarily
the correct ones.

We now update _output_dir in exclude_execution() so the execution item
instances return the correct output resources.