Mark deleted and renamed files as missing references
Instead of removing or renaming Data connection's file references behind
user's back, we now mark deleted or renamed files as "missing" in Data connection.
This fixes the problem that we essentially modify the project without
using the undo stack and thus not showing the user that something had changed.

In case of file rename, we now wait for 1s before trying to refresh the reference.
This is a workaround to e.g. Excel files that get renamed and then deleted
when the files are saved in Excel resulting in Data connection losing the
reference. In case 1s is not enough (file on network drive, for example), users
can manually fix the problem by forcing Data connection to refresh the missing
reference.

Re Spine-project/Spine-Toolbox#1691