fix tool spec editor having same file in two places
The tool spec editor no longer can have the same file as the main program file as well as an additional program file. If the user tries to add a file that is already loaded into the tool spex editor, the file will be skipped and a popup will be shown.

Re spine-tools\Spine-Toolbox#2128