Speed up checking Tools' notifications (#143)
On large projects that have a large amount of files as possible resources for a Tool, checking the Tool's notifications took a long time. Now the checks for duplicate files in resource files and checking if the input file requirements are met are only conducted when either the requirements or resources have changed. The checking of file duplicates has also been made faster.

Re spine-tools/Spine-Toolbox#2083