Prevent FileNotFoundError crashing Exporter
Looks like in some circumstances, Exporter's _export_to_file() may
crash with a Traceback when creating the output directory when the
output directory is on OneDrive. We now expect the FileNotFoundError
and log it instead of crashing do_work() which would result in
Toolbox erroneously thinking that the Exporter never finished.

Re spine-tools/Spine-Toolbox#2211