Improvements for executable tool spec editor
- Shel types that are not supported on the os are now disabled
instead of letting the user select them and showing an error.
Bash is disabled for windows and cmd and powershell are disabled
for non windows platforms. Added tooltips to disabled items
stating the incompatibility between the shell and current os.
- When executable tool is selected and there is a main
program file, the Command -line-edit and the Shell -combobox
are disabled since they are not used in that case anyhow.

Re spine-tools/Spine-Toolbox#2205