Warn about uncommitted changes with Data Store notification
Now when a db map has uncommitted changes, all Data Stores
that share that same db map will show a notification about
the dirtiness of that Data Store. Edge cases that should
now be covered (and come to mind...):

- If more than one DS points to the same db map, all DSs
pointing to that should be notified if the db map is dirty.
- Everything should still work after renaming a DS

Re spine-tools/Spine-Toolbox#1219