Rewrite Exporter's preview updater using multiprocessing
PreviewUpdater was using QThreads to write the preview tables
which made the Specification editor unresponsive especially when
dealing with large number of tables. Also, it could deadlock when
multiple threads accessed the same database via multiple
DatabaseMappings (this is actually a spinedb_api issue).

This replaces the QThreads by a single worker process.
Because the tables are written in a separate process, we don't
encounter Python's global interpreter lock.

Re #2898