Make URL dicts more explicit; make port an int
The URL dicts used internally by URL selector widget, Data Connection
and Data Store are now TypedDicts called UrlDict,
and functions that use them have been typed accordingly.
This hopefully makes it clearer when to use UrlDict, string or
SQLAlchemy URL --- we use all three in different places for
different purposes.

The "port" field in UrlDict is now a proper integer,
it used to be a string.

Re spine-tools/Spine-Toolbox#3178